requirements_dev.txt
cryptography==41.0.7
mysql-connector-python<=8.0.29
fastapi==0.101.0
httpx==0.23.3
importlib_resources==5.8.0; python_version<'3.9'
ispyb==6.11.0
jinja2==3.1.3
numpy==1.25
procrunner==2.3.3
prometheus_client==0.17.1
pillow==10.2.0
pydantic==1.10.1
pytest-cov==4.0.0
pytest==7.1.2
pyyaml==6.0
rich==13.3.3
starlette==0.27.0
sqlmodel==0.0.8
textual==0.42.0
uvicorn[standard]==0.18.2
websocket-client==1.3.3
werkzeug==3.0.1
xmltodict==0.13.0
zocalo==0.21.0
